+++ example/lib/feature/features/home/view/freature_home_page.dart
@@ -1,0 +1,20 @@
+// features/home/view/home_page.dart
+import 'package:flutter/material.dart';
+import 'package:rxflare/rxflare.dart';
+import '../controller/home_controller.dart';
+import 'home_view.dart';
+
+class FreatureHomePage extends StatelessWidget {
+  const FreatureHomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return RxParent<HomeController>(
+      name: "home",                    // 重要：多页面隔离
+      dependency: HomeController(),
+      child: const HomeView(),
+    );
+  }
+}
+
+